requirements.txt
Flask
Flask-Login
Flask-WTF
WTForms
PyPDF2
python-docx
reportlab
transformers
nltk
scikit-learn
python-dotenv
Werkzeug
email-validator
Flask-SQLAlchemy
SQLAlchemy
bcrypt
pdfkit
google-generativeai>=0.3.0
sentence-transformers
spacy
python-dotenv>=1.0.0